requirements.txt
# Voice interview (run_questions.py)
elevenlabs>=1.0.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.0
numpy>=1.24.0
